Langchain_1/requirements.txt
fastapi
uvicorn
python-dotenv
langchain_openai
langchain 
langchain-google-firestore 
firestore 
langchain_google_vertexai
sentence_transformers 
pinecone
langchain_pinecone
langchain_huggingface
PyPDF2
pymupdf
#pip install -r requirements.txt